+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## [1.2.0] - 2021-07-26
 
 ### Changed